+++ zetaclient/chains/sui/signer/signer_tx.go
@@ -2,6 +2,7 @@ package signer
 
 import (
 	"context"
+	"encoding/hex"
 	"strconv"
 
 	"github.com/block-vision/sui-go-sdk/models"
@@ -14,10 +15,16 @@ import (
 
 const funcWithdraw = "withdraw"
 
+// TODO: use these functions in PTB building
+// https://github.com/zeta-chain/node/issues/3741
+//const funcWithdrawImpl = "withdraw_impl"
+//const funcOnCall = "on_call"
+
 // buildWithdrawal builds unsigned withdrawal transaction using CCTX and Sui RPC
 // https://github.com/zeta-chain/protocol-contracts-sui/blob/0245ad3a2eb4001381625070fd76c87c165589b2/sources/gateway.move#L117
 func (s *Signer) buildWithdrawal(ctx context.Context, cctx *cctypes.CrossChainTx) (tx models.TxnMetaData, err error) {
 	params := cctx.GetCurrentOutboundParam()
+
 	coinType := ""
 
 	// Basic common-sense validation & coin-type determination
@@ -28,8 +35,6 @@ func (s *Signer) buildWithdrawal(ctx context.Context, cctx *cctypes.CrossChainTx
 		return tx, errors.Errorf("invalid protocol version %q", cctx.ProtocolContractVersion)
 	case cctx.InboundParams == nil:
 		return tx, errors.New("inbound params are nil")
-	case cctx.IsWithdrawAndCall():
-		return tx, errors.New("withdrawAndCall is not supported yet")
 	case params.CoinType == coin.CoinType_Gas:
 		coinType = string(sui.SUI)
 	case params.CoinType == coin.CoinType_ERC20:
@@ -39,12 +44,6 @@ func (s *Signer) buildWithdrawal(ctx context.Context, cctx *cctypes.CrossChainTx
 		return tx, errors.Errorf("unsupported coin type %q", params.CoinType.String())
 	}
 
-	var (
-		nonce     = strconv.FormatUint(params.TssNonce, 10)
-		recipient = params.Receiver
-		amount    = params.Amount.String()
-	)
-
 	// Gas budget is gas limit * gas price
 	gasPrice, err := strconv.ParseUint(params.GasPrice, 10, 64)
 	if err != nil {
@@ -52,10 +51,32 @@ func (s *Signer) buildWithdrawal(ctx context.Context, cctx *cctypes.CrossChainTx
 	}
 	gasBudget := strconv.FormatUint(gasPrice*params.CallOptions.GasLimit, 10)
 
+	// Retrieve withdraw cap ID
 	withdrawCapID, err := s.getWithdrawCapIDCached(ctx)
 	if err != nil {
 		return tx, errors.Wrap(err, "unable to get withdraw cap ID")
 	}
+
+	// build tx depending on the type of transaction
+	if cctx.IsWithdrawAndCall() {
+		return s.buildWithdrawAndCallTx(ctx, params, coinType, gasBudget, withdrawCapID, cctx.RelayedMessage)
+	}
+	return s.buildWithdrawTx(ctx, params, coinType, gasBudget, withdrawCapID)
+}
+
+// buildWithdrawTx builds unsigned withdraw transaction
+func (s *Signer) buildWithdrawTx(
+	ctx context.Context,
+	params *cctypes.OutboundParams,
+	coinType,
+	gasBudget,
+	withdrawCapID string,
+) (models.TxnMetaData, error) {
+	var (
+		nonce     = strconv.FormatUint(params.TssNonce, 10)
+		recipient = params.Receiver
+		amount    = params.Amount.String()
+	)
 
 	req := models.MoveCallRequest{
 		Signer:          s.TSS().PubKey().AddressSui(),
@@ -68,6 +89,51 @@ func (s *Signer) buildWithdrawal(ctx context.Context, cctx *cctypes.CrossChainTx
 	}
 
 	return s.client.MoveCall(ctx, req)
+}
+
+// buildWithdrawAndCallTx builds unsigned withdrawAndCall
+// a withdrawAndCall is a PTB transaction that contains a withdraw_impl call and a on_call call
+func (s *Signer) buildWithdrawAndCallTx(
+	ctx context.Context,
+	params *cctypes.OutboundParams,
+	coinType,
+	gasBudget,
+	withdrawCapID,
+	payload string,
+) (models.TxnMetaData, error) {
+	// decode and parse the payload to object the on_call arguments
+	payloadBytes, err := hex.DecodeString(payload)
+	if err != nil {
+		return models.TxnMetaData{}, errors.Wrap(err, "unable to decode payload hex bytes")
+	}
+
+	var cp sui.CallPayload
+	if err := cp.UnpackABI(payloadBytes); err != nil {
+		return models.TxnMetaData{}, errors.Wrap(err, "unable to parse withdrawAndCall payload")
+	}
+
+	// Note: logs not formatted in standard, it's a temporary log
+	s.Logger().Std.Info().Msgf(
+		"WithdrawAndCall called with type arguments %v, object IDs %v, message %v",
+		cp.TypeArgs,
+		cp.ObjectIDs,
+		cp.Message,
+	)
+
+	// keep lint quiet without using _ in params
+	_ = ctx
+	_ = params
+	_ = coinType
+	_ = gasBudget
+	_ = withdrawCapID
+
+	// TODO: check all object IDs are share object here
+	// https://github.com/zeta-chain/node/issues/3755
+
+	// TODO: build PTB here
+	// https://github.com/zeta-chain/node/issues/3741
+
+	return models.TxnMetaData{}, errors.New("not implemented")
 }
 
 // broadcast attaches signature to tx and broadcasts it to Sui network. Returns tx digest.