+++ changelog.md
@@ -2,6 +2,10 @@
 
 # UNRELEASED
 
+### Breaking Changes
+
+* All Solana inbounds have new optional param `revert_options`. Solana Gateway program must be upgraded after node upgrade.
+
 ### Features
 
 * [3672](https://github.com/zeta-chain/node/pull/3672) - zetaclient: cache tss signatures for performance.
@@ -10,14 +14,18 @@
 * [3710](https://github.com/zeta-chain/node/pull/3710) - support preflight RPC health metrics before fully enable a chain
 * [3377](https://github.com/zeta-chain/node/pull/3377) - have zetacore increase the gas price in pending Bitcoin cctxs and burns additional gas fees
 * [3750](https://github.com/zeta-chain/node/pull/3750) - support simple call from solana
+* [3764](https://github.com/zeta-chain/node/pull/3764) - add payload parsing for Sui WithdrawAndCall
+* [3756](https://github.com/zeta-chain/node/pull/3756) - parse revert options in solana inbounds
 
 ### Refactor
 
 * [3709](https://github.com/zeta-chain/node/pull/3709) - improve cctx error message for out of gas errors when creating outbound
+* [3777](https://github.com/zeta-chain/node/pull/3777) - use SignBatch keysign for solana outbound tx and fallback tx
 
 ### Fixes
 * [3711](https://github.com/zeta-chain/node/pull/3711) - fix TON call_data parsing
 * [3717](https://github.com/zeta-chain/node/pull/3717) - fix solana withdraw and call panic
+* [3770](https://github.com/zeta-chain/node/pull/3770) - improve fallback tx error handling
 
 ### Tests